修改bug

+++ app/src/main/java/com/novel/cn/mvp/model/entity/ChargeChapter.kt
@@ -1,3 +1,3 @@
 package com.novel.cn.mvp.model.entity
 
-data class ChargeChapter(val isSubscibe:Boolean,val userId:String,val goldNumber:Int)
+data class ChargeChapter(val isSubscibe:Boolean,val userId:String,val goldNumber:Float)
