+++ app/src/main/java/com/novel/cn/mvp/model/entity/UserAccountBean.kt
@@ -3,7 +3,7 @@ package com.novel.cn.mvp.model.entity
 data class UserAccountBean(
     val diamondNumber: Int,
     val diamondValidity: Any,
-    val goldNumber: Int,
+    val goldNumber: Float,
     val isSubscibe: Any,
     val monthRecommendNumber: Int,
     val monthRecommendValidity: String,
